// lib/api/generated/models/exerciseLibraryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 1.1.0.22+23
 */

/**
 * Biblioteka ćwiczeń klubowa (082).
 */
export interface ExerciseLibraryItem {
  created_at: string;
  created_by: string;
  /** @nullable */
  default_reps?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  default_sets?: number | null;
  id: string;
  name: string;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
  updated_at: string;
  /** @nullable */
  video_url?: string | null;
}